# lib/api-zod/src/generated/types/mhraDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MhraDocument {
  title: string;
  productName: string;
  plNumber?: string[];
  substanceName?: string[];
  docType: string;
  territory?: string | null;
  documentUrl: string;
  fileName?: string;
  fileSize?: number;
  created?: string;
}