# lib/api-zod/src/generated/types/mhraSearchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MhraDocument } from "./mhraDocument";

export interface MhraSearchResponse {
  results: MhraDocument[];
  totalCount: number;
  page: number;
  pageSize: number;
  query: string;
}